Ensure order is properly updated after payment change.
This is a cherrypick from solidus/solidus@a1dd8582. Although that
commit's goals is better encapsulation, our need here is driven from
the fact that after voiding a payment the order payment_state and
payment_total is not updated.

Although the existing Spree code does update those two fields, because
of the selective use of OrderUpdater it doesn't actually persist the
values to the DB because it only does that if the payment is complete.
Obviously a bug but rather than fix that bug we just follow the example
of Solidus and let the OrderUpdater be responsible for updating the
order. It does it correctly leaving this code simpler.